Refactor Java `int` and `long` validation logic.
- Replaced `Integer#MIN_VALUE` with `Config#INVALID_INT` for
consistency.
- Replaced `Long#MIN_VALUE` with `Config#INVALID_LONG` for consistency.
- Updated checks to use `F#isPresent` and `F#isNotPresent` for
validation.

+++ symja_android_library/matheclipse-io/src/main/java/tech/tablesaw/columns/datetimes/PackedLocalDateTime.java
@@ -1,23 +1,18 @@
 /*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  */
 
 package tech.tablesaw.columns.datetimes;
 
 import static tech.tablesaw.columns.datetimes.DateTimeColumnType.missingValueIndicator;
-
-import com.google.common.base.Strings;
-import com.google.common.primitives.Ints;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -28,6 +23,8 @@ import java.time.temporal.ChronoField;
 import java.time.temporal.TemporalField;
 import java.time.temporal.WeekFields;
 import java.util.Locale;
+import com.google.common.base.Strings;
+import com.google.common.primitives.Ints;
 import tech.tablesaw.columns.dates.PackedLocalDate;
 import tech.tablesaw.columns.instant.PackedInstant;
 import tech.tablesaw.columns.times.PackedLocalTime;
@@ -42,13 +39,16 @@ import tech.tablesaw.columns.times.PackedLocalTime;
  * A short localdatetime packed into a single long value. The long is comprised of an int for the
  * date and an int for the time
  *
- * <p>The bytes are packed into the date int as: First two bytes: short (year) next byte (month of
+ * <p>
+ * The bytes are packed into the date int as: First two bytes: short (year) next byte (month of
  * year) last byte (day of month)
  *
- * <p>The bytes are packed into the time int as First byte: hourOfDay next byte: minuteOfHour last
- * two bytes (short): millisecond of minute
+ * <p>
+ * The bytes are packed into the time int as First byte: hourOfDay next byte: minuteOfHour last two
+ * bytes (short): millisecond of minute
  *
- * <p>Storing the millisecond of minute in an short requires that we treat the short as if it were
+ * <p>
+ * Storing the millisecond of minute in an short requires that we treat the short as if it were
  * unsigned. Unfortunately, Neither Java nor Guava provide unsigned short support so we use char,
  * which is a 16-bit unsigned int to store values of up to 60,000 milliseconds (60 secs * 1000)
  */
@@ -124,26 +124,18 @@ public class PackedLocalDateTime extends PackedInstant {
     int date = date(dateTime);
     int time = time(dateTime);
 
-    return ""
-        + PackedLocalDate.getYear(date)
-        + "-"
-        + Strings.padStart(Byte.toString(PackedLocalDate.getMonthValue(date)), 2, '0')
-        + "-"
-        + Strings.padStart(Byte.toString(PackedLocalDate.getDayOfMonth(date)), 2, '0')
-        + "T"
-        + Strings.padStart(Byte.toString(PackedLocalTime.getHour(time)), 2, '0')
-        + ":"
-        + Strings.padStart(Byte.toString(PackedLocalTime.getMinute(time)), 2, '0')
-        + ":"
-        + Strings.padStart(Byte.toString(PackedLocalTime.getSecond(time)), 2, '0')
-        + "."
+    return "" + PackedLocalDate.getYear(date) + "-"
+        + Strings.padStart(Byte.toString(PackedLocalDate.getMonthValue(date)), 2, '0') + "-"
+        + Strings.padStart(Byte.toString(PackedLocalDate.getDayOfMonth(date)), 2, '0') + "T"
+        + Strings.padStart(Byte.toString(PackedLocalTime.getHour(time)), 2, '0') + ":"
+        + Strings.padStart(Byte.toString(PackedLocalTime.getMinute(time)), 2, '0') + ":"
+        + Strings.padStart(Byte.toString(PackedLocalTime.getSecond(time)), 2, '0') + "."
         + Strings.padStart(String.valueOf(PackedLocalTime.getMilliseconds(time)), 3, '0');
   }
 
   public static int getDayOfYear(long packedDateTime) {
     return getMonth(packedDateTime).firstDayOfYear(isLeapYear(packedDateTime))
-        + getDayOfMonth(packedDateTime)
-        - 1;
+        + getDayOfMonth(packedDateTime) - 1;
   }
 
   public static int getWeekOfYear(long packedDateTime) {
@@ -156,7 +148,8 @@ public class PackedLocalDateTime extends PackedInstant {
   }
 
   public static boolean isLeapYear(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     return IsoChronology.INSTANCE.isLeapYear(getYear(packedDateTime));
   }
 
@@ -195,25 +188,29 @@ public class PackedLocalDateTime extends PackedInstant {
   }
 
   public static boolean isInQ1(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     Month month = getMonth(packedDateTime);
     return month == Month.JANUARY || month == Month.FEBRUARY || month == Month.MARCH;
   }
 
   public static boolean isInQ2(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     Month month = getMonth(packedDateTime);
     return month == Month.APRIL || month == Month.MAY || month == Month.JUNE;
   }
 
   public static boolean isInQ3(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     Month month = getMonth(packedDateTime);
     return month == Month.JULY || month == Month.AUGUST || month == Month.SEPTEMBER;
   }
 
   public static boolean isInQ4(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     Month month = getMonth(packedDateTime);
     return month == Month.OCTOBER || month == Month.NOVEMBER || month == Month.DECEMBER;
   }
@@ -227,43 +224,50 @@ public class PackedLocalDateTime extends PackedInstant {
   }
 
   public static boolean isSunday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.SUNDAY;
   }
 
   public static boolean isMonday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.MONDAY;
   }
 
   public static boolean isTuesday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.TUESDAY;
   }
 
   public static boolean isWednesday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.WEDNESDAY;
   }
 
   public static boolean isThursday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.THURSDAY;
   }
 
   public static boolean isFriday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.FRIDAY;
   }
 
   public static boolean isSaturday(long packedDateTime) {
-    if (packedDateTime == missingValueIndicator()) return false;
+    if (packedDateTime == missingValueIndicator())
+      return false;
     DayOfWeek dayOfWeek = getDayOfWeek(packedDateTime);
     return dayOfWeek == DayOfWeek.SATURDAY;
   }
@@ -392,7 +396,7 @@ public class PackedLocalDateTime extends PackedInstant {
     if (localDateTime == null) {
       throw new IllegalArgumentException("Cannot get millisecond of day for missing value");
     }
-    long total = (long) localDateTime.get(ChronoField.MILLI_OF_SECOND);
+    long total = localDateTime.get(ChronoField.MILLI_OF_SECOND);
     total += localDateTime.getSecond() * 1000;
     total += localDateTime.getMinute() * 60 * 1000;
     total += localDateTime.getHour() * 60 * 60 * 1000;
